Move page models
Move page models to their own namespace away from Pages.
