+++ tests/Dashboard.Tests/DashboardTests.cs
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 
 using System.Runtime.CompilerServices;
-using MartinCostello.Benchmarks.Pages;
+using MartinCostello.Benchmarks.PageModels;
 using Microsoft.AspNetCore.WebUtilities;
 using Microsoft.Playwright;
