+++ tests/Dashboard.Tests/PageModels/TokenPage.cs
@@ -3,7 +3,7 @@
 
 using Microsoft.Playwright;
 
-namespace MartinCostello.Benchmarks.Pages;
+namespace MartinCostello.Benchmarks.PageModels;
 
 public sealed class TokenPage(IPage page) : AppPage(page)
 {
